카운터 리팩토링
관심사에 맡게 컴포넌트를 분리함.

App 컴포넌트는 count 상태를 변경시키는 비즈니스 로직에만 관심을 가지고
뷰에 대한 부분은 Page 컴포넌트에 일임한다.

Page 컴포넌트는 카운터앱을 구성하는 컴포넌트를 가지고 있다.

Button 컴포넌트는 props가 주어진대로 UI를 그리고 이벤트를 실행할 뿐이다.

Buttons 컴포넌트는 요구사항대로 5개의 카운터 버튼을 그리고 비즈니스 로직을 전달한다.
